fg_log: Add basic implementation of 'fg_log'
The current implementation of fg_log lists the commits using the
'--online' option of 'git log'. The preview shows a more detailled
version of the commit using the '--stat' option of 'git log'.

If a commit has been selected, the log between this commit and the
current HEAD is shown using the '--stat' option.