refac: Add flag to enable exprimental preview #wip
The flag `fg_var_previewUseSed` can be set to false in order not to use
`sed` in the preview commands. Instead a `fzf` pattern will be used.

This was done to simplify preview commands. This is not considered to be
final. Some cases like a renamed file are not handeld correctly using
fzf patterns in the preview.